Fix types in DeviceSelector
